cloud_function_vision_api/requirements.txt
google-cloud-vision
google-cloud-storage
os
tempfile